chore: reorganized routes, unifying students and classes with school config (and renaming to school management)
