+++ lib/lanttern_web/live/pages/school/cycles_component.ex
@@ -1,4 +1,4 @@
-defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
+defmodule LantternWeb.SchoolLive.CyclesComponent do
   use LantternWeb, :live_component
 
   alias Lanttern.Schools
@@ -12,11 +12,7 @@ defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
     ~H"""
     <div>
       <div :if={@is_school_manager} class="flex justify-end gap-6 p-4">
-        <.action
-          type="link"
-          patch={~p"/school_config/cycles?new=true"}
-          icon_name="hero-plus-circle-mini"
-        >
+        <.action type="link" patch={~p"/school/cycles?new=true"} icon_name="hero-plus-circle-mini">
           <%= gettext("Add cycle") %>
         </.action>
       </div>
@@ -54,7 +50,7 @@ defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
                 size="sm"
                 theme="ghost"
                 rounded
-                patch={~p"/school_config/cycles?edit=#{cycle.id}"}
+                patch={~p"/school/cycles?edit=#{cycle.id}"}
               />
             </:action>
           </.data_grid>
@@ -72,7 +68,7 @@ defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
         id="cycle-form-overlay"
         cycle={@cycle}
         title={@cycle_form_overlay_title}
-        on_cancel={JS.patch(~p"/school_config/cycles")}
+        on_cancel={JS.patch(~p"/school/cycles")}
         notify_component={@myself}
       />
     </div>
@@ -90,7 +86,7 @@ defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
   def update(%{action: {CycleFormOverlayComponent, {:created, _cycle}}}, socket) do
     nav_opts = [
       put_flash: {:info, gettext("Cycle created successfully")},
-      push_navigate: [to: ~p"/school_config/cycles"]
+      push_navigate: [to: ~p"/school/cycles"]
     ]
 
     {:ok, delegate_navigation(socket, nav_opts)}
@@ -99,7 +95,7 @@ defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
   def update(%{action: {CycleFormOverlayComponent, {:updated, cycle}}}, socket) do
     nav_opts = [
       put_flash: {:info, gettext("Cycle updated successfully")},
-      push_navigate: [to: ~p"/school_config/cycles"]
+      push_navigate: [to: ~p"/school/cycles"]
     ]
 
     socket =
@@ -113,7 +109,7 @@ defmodule LantternWeb.SchoolConfigLive.CyclesComponent do
   def update(%{action: {CycleFormOverlayComponent, {:deleted, cycle}}}, socket) do
     nav_opts = [
       put_flash: {:info, gettext("Cycle deleted successfully")},
-      push_patch: [to: ~p"/school_config/cycles"]
+      push_patch: [to: ~p"/school/cycles"]
     ]
 
     socket =
