+++ lib/lanttern_web/live/pages/school/school_live.html.heex
@@ -2,7 +2,7 @@
   <:title><%= @current_user.current_profile.school_name %></:title>
   <div class="px-4">
     <.neo_tabs>
-      <:tab patch={~p"/school/classes"} is_current={@live_action in [:show, :manage_classes]}>
+      <:tab patch={~p"/school/classes"} is_current={@live_action == :manage_classes}>
         <%= if @current_user.current_profile.current_school_cycle,
           do: "#{@current_user.current_profile.current_school_cycle.name} #{gettext("classes")}",
           else: gettext("Classes") %>
@@ -10,11 +10,20 @@
       <:tab patch={~p"/school/students"} is_current={@live_action == :manage_students}>
         <%= gettext("All students") %>
       </:tab>
+      <:tab patch={~p"/school/cycles"} is_current={@live_action == :manage_cycles}>
+        <%= gettext("Cycles") %>
+      </:tab>
+      <:tab
+        patch={~p"/school/moment_cards_templates"}
+        is_current={@live_action == :manage_moment_cards_templates}
+      >
+        <%= gettext("Templates") %>
+      </:tab>
     </.neo_tabs>
   </div>
 </.header_nav>
 <.live_component
-  :if={@live_action in [:show, :manage_classes]}
+  :if={@live_action == :manage_classes}
   module={ClassesComponent}
   id={:school_classes}
   current_user={@current_user}
@@ -29,4 +38,20 @@
   params={@params}
   live_action={@live_action}
   is_school_manager={@is_school_manager}
+/>
+<.live_component
+  :if={@live_action == :manage_cycles}
+  module={CyclesComponent}
+  id={:school_cycles}
+  current_user={@current_user}
+  params={@params}
+  is_school_manager={@is_school_manager}
+/>
+<.live_component
+  :if={@live_action == :manage_moment_cards_templates}
+  module={MomentCardsTemplatesComponent}
+  id={:school_moment_cards_templates}
+  current_user={@current_user}
+  params={@params}
+  is_content_manager={@is_content_manager}
 />
