+++ lib/lanttern_web/live/pages/school/students/id/student_live.ex
@@ -91,7 +91,7 @@ defmodule LantternWeb.StudentLive do
     socket =
       socket
       |> put_flash(:info, gettext("Student deleted successfully"))
-      |> push_navigate(to: ~p"/school")
+      |> push_navigate(to: ~p"/school/classes")
 
     {:noreply, socket}
   end
