+++ lib/lanttern_web/live/pages/school/students/id/student_live.html.heex
@@ -1,6 +1,6 @@
 <div>
   <.header_nav current_user={@current_user}>
-    <:breadcrumb navigate={~p"/school"}><%= gettext("School") %></:breadcrumb>
+    <:breadcrumb navigate={~p"/school/classes"}><%= gettext("School") %></:breadcrumb>
     <:title><%= @student.name %></:title>
     <div class="flex items-center justify-between gap-4 px-4">
       <.neo_tabs id="student-nav-tabs">
